+++ src/dribik/__init__.py
@@ -2,5 +2,5 @@
 
 from __future__ import annotations
 
-__version__ = "0.0.2-beta"
+__version__ = "0.1.0-beta"
 __author__ = "Dribik Project"